mv: numba test to non-numba test. Add abstract class check
